requirements.txt
APScheduler==3.9.1
aiohttp==3.8.2
python-dotenv==0.21.0
aiogram==2.22.1
aioredis==2.0.1
orjson==3.8.0
pydantic==1.10.2
pandas==1.5.0
xlrd==2.0.1
numpy==1.23.3
sqlalchemy==1.4.41
asyncpg==0.26.0
openpyxl==3.0.10
